refactor(tui): split ui into a separate mod for app
